Update egui to version 0.27.0
Seems like there is a change in the behaviour of clicking/dragging
windows - clicking the node and dragging it still selects the window,
but not longer drags it.

Also seems like the size of the window changes very slightly when it is
selected causing a bit of jitter in the edges, would be good to solve
this before landing.